requirements.txt
Django
djangorestframework
django-environ
djlint
django-cors-headers

whitenoise
Pillow
gunicorn
uvicorn[standard]
uvicorn-worker
